machine_learning/requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
